% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris2}
\alias{iris2}
\title{Edgar Anderson's Iris Data}
\format{
\code{"iris"} is a data frame with 150 cases (rows) and 5 variables
(columns) named Sepal.Length, Sepal.Width, Petal.Length, Petal.Width, and Species.
}
\source{
\href{https://phenome.jax.org/projects/Moore1b}{QTL Archive}
}
\usage{
data(iris2)
}
\description{
This famous (Fisher's or Anderson's) iris data set gives the measurements in
centimeters of the variables sepal length and width and petal length and width,
respectively, for 50 flowers from each of 3 species of iris. The species are
Iris setosa, versicolor, and virginica.
}
\examples{
data(iris2)
plot_dist(dat = iris2, var_name = "Petal.Width")
}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988)
The New S Language. Wadsworth & Brooks/Cole.
}
\keyword{datasets}
