% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist.R
\name{plot_dist}
\alias{plot_dist}
\title{Plot Histogram}
\usage{
plot_dist(dat, var_name)
}
\arguments{
\item{dat}{a data.frame containing the data to plot.}

\item{var_name}{a character vector of the column name in the data.}
}
\value{
a histogram plot.
}
\description{
Plot the histogram of a numeric variable within a data frame
}
\examples{
plot_dist(dat = iris, var_name = "Petal.Length")

}
